requirements.txt
beautifulsoup4==4.9.3
requests==2.26.0
newspaper3k==0.2.8
pygooglenews==0.1.2
pandas==2.1.0
pymongo==4.5.0
fake-useragent == 1.3.0